fix(docz-theme-default): codemirror line end bug
