+++ packages/docz-theme-default/package.json
@@ -35,7 +35,7 @@
     "prop-types": "15.6.2",
     "re-resizable": "^4.8.1",
     "react": "^16.4.2",
-    "react-codemirror": "^1.0.0",
+    "react-codemirror2": "^5.1.0",
     "react-dom": "^16.4.2",
     "react-emotion": "^9.2.8",
     "react-feather": "^1.1.1",
