+++ packages/docz-theme-default/src/types.d.ts
@@ -5,7 +5,7 @@ declare module 'codemirror/mode/javascript/javascript'
 declare module 'codemirror/mode/jsx/jsx'
 declare module 'codemirror/mode/css/css'
 declare module 'codemirror/addon/edit/matchbrackets'
-declare module 'react-codemirror'
+declare module 'react-codemirror2'
 declare module 'react-feather/dist/icons/edit-2'
 declare module 'react-feather/dist/icons/chevron-down'
 declare module 'react-feather/dist/icons/search'
